+++ app/src/main/java/com/mapbox/maps/testapp/examples/markersandcallouts/viewannotation/ViewAnnotationShowcaseActivity.kt
@@ -81,7 +81,8 @@ class ViewAnnotationShowcaseActivity :
         binding.fabStyleToggle.setOnClickListener {
           when (style?.styleURI) {
             Style.STANDARD -> loadStyle(prepareStyle(Style.SATELLITE_STREETS, bitmap))
-            Style.SATELLITE_STREETS -> loadStyle(prepareStyle(Style.STANDARD, bitmap))
+            Style.SATELLITE_STREETS -> loadStyle(prepareStyle(Style.STANDARD_SATELLITE, bitmap))
+            Style.STANDARD_SATELLITE -> loadStyle(prepareStyle(Style.STANDARD, bitmap))
           }
         }
         Toast.makeText(this@ViewAnnotationShowcaseActivity, STARTUP_TEXT, Toast.LENGTH_LONG).show()